% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHO_lambda.R
\name{CHO_lambda}
\alias{CHO_lambda}
\title{CHO_lambda}
\usage{
CHO_lambda(fdata_all)
}
\arguments{
\item{fdata_all}{A list of GWAS summary statistics for E populations,
each element is a list consists of betaGX, sebetaGX, betaGY and sebetaGY.
betaGX is nSNP*nexposure matrix of G-X association (beta coefficient),
sebetaGX is the corresponding standard error matrix,
betaGY is vector of G-Y association (beta coefficient),
sebetaGY is the corresponding standard error vector.}
}
\value{
the ridge plot.
}
\description{
plot ridge plot for choosing lambda.
}
